Created BinPyBits class inheriting from BitArray. Moved bittools.py to bittools directory.
